move to another end2end test file
